[修改] 新增嵌入式菜单标识

+++ src/main/java/com/microtomato/hirun/modules/user/entity/po/MenuRole.java
@@ -46,7 +46,7 @@ public class MenuRole extends BaseEntity {
      * 菜单ID
      */
     @TableField("MENU_ID")
-    private Integer menuId;
+    private Long menuId;
 
     /**
      * 是否有效（0:有效）
